#2 Fix Topic Subscribing and Handling
This commit modifies the 'on connect' event listener to properly subscribe to the relevant topics. All of the topics are now handled in a switch-function inside the 'on message' event listener. The use of 'reqId' has now been made consistent in only using the mongo _id, which will be refactored to be automatically fetched on successful login in a future commit. Furthermore, some minor changes for the purpose of readability and code robustness have been made.